requirements.txt
Flask ==1.1.2
flask-restx==0.2.0
flask-restplus==0.13.0
Werkzeug==1.0.1
requests==2.24.0
PyJWT==1.7.1